Added Model.timex(), labelx() and binding(). Deprecated similar lib.multi methods. Closes #394
